ml/requirements.txt
# Core ML Libraries
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning
tensorflow>=2.14.0
torch>=2.1.0
torchvision>=0.16.0

# Time Series Analysis
statsmodels>=0.14.0
prophet>=1.1.4
pmdarima>=2.0.4

# Data Processing
polars>=0.19.0
pyarrow>=14.0.0
dask>=2023.10.0

# Feature Engineering
feature-engine>=1.6.0
category-encoders>=2.6.0

# Model Management
mlflow>=2.8.0
wandb>=0.16.0
optuna>=3.4.0

# Model Serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.2

# Data Validation
great-expectations>=0.18.0
evidently>=0.4.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
bokeh>=3.3.0

# Jupyter & Notebooks
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.1.0
nbconvert>=7.10.0

# Database Connectivity
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
redis>=5.0.1

# Cloud Storage
boto3>=1.29.3
azure-storage-blob>=12.19.0
google-cloud-storage>=2.10.0

# Streaming & Message Queues
kafka-python>=2.0.2
apache-airflow>=2.7.0

# Monitoring
prometheus-client>=0.17.1

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
click>=8.1.7
tqdm>=4.66.0
joblib>=1.3.0

# Development & Testing
pytest>=7.4.3
pytest-cov>=4.1.0
black>=23.10.1
flake8>=6.1.0
mypy>=1.6.1
isort>=5.12.0

# Documentation
mkdocs>=1.5.3
mkdocs-material>=9.4.0 